requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-admin-rangefilter==0.5.1
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.1
Pillow==6.2.1
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0